Convert extra panel index from u8 to usize

+++ src/components/extra_tab.rs
@@ -2,7 +2,7 @@ use crate::ExtraCategory;
 use yew::{html, Callback, Children, Component, Context, Html, Properties};
 
 pub enum Msg {
-    Click(u8),
+    Click(usize),
 }
 
 #[derive(Properties, PartialEq)]
@@ -14,7 +14,7 @@ pub struct Props {
 }
 
 pub struct ExtraTab {
-    active_panel: u8,
+    active_panel: usize,
 }
 
 impl Component for ExtraTab {
@@ -38,7 +38,7 @@ impl Component for ExtraTab {
         let link = ctx.link();
         let props = ctx.props();
 
-        let category = props.categories[self.active_panel as usize];
+        let category = props.categories[self.active_panel];
         let on_clear = props.on_clear.reform(move |_| category);
 
         let iter = props.names.iter().zip(props.children.iter()).zip(0..);
